Buttons logic and few bugs were edited
1. All project defines were added to project options window. User must add or/and delete defines if he needed.
MEM_PROT_DIS - if enabled memory won't be protected after downloading firmware to MCU. If disabled user must clear ROP option byte and all flash and epprom data will be erased after this action.
2. HIGH_STATE_START - if enabled relay starts with high state. I f disabled it's vice versa.
3. Variable size (uint8_t) was added to private defines in indicator.c
4. There was a bug with enter menu. isHighStateTime flag was set wrongly in init function. Also high state dot was incorrectly set
5. Buttons logic in enter menu was changed. Long set button press just starts work execution without saving data. If user wants to save in EEPROM plus and minus buttons must be pressed together for at least one second. All data will be saved and enter mode will be executed after.
6. Clear plus and minus button functions were added to work mode.

Addtional:
1. In enter menu high state data are shown with dot at the right indicator.

+++ Project/user/src/indicator.c
@@ -2,14 +2,14 @@
 
 //Private defines
 enum{
-  IND_A_POS = (1 << 0),
-  IND_B_POS = (1 << 1),
-  IND_C_POS = (1 << 2),
-  IND_D_POS = (1 << 3),
-  IND_E_POS = (1 << 4),
-  IND_F_POS = (1 << 5),
-  IND_G_POS = (1 << 6),
-  IND_DP_POS = (1 << 7)
+  IND_A_POS = ((uint8_t)(1 << 0)),
+  IND_B_POS = ((uint8_t)(1 << 1)),
+  IND_C_POS = ((uint8_t)(1 << 2)),
+  IND_D_POS = ((uint8_t)(1 << 3)),
+  IND_E_POS = ((uint8_t)(1 << 4)),
+  IND_F_POS = ((uint8_t)(1 << 5)),
+  IND_G_POS = ((uint8_t)(1 << 6)),
+  IND_DP_POS = ((uint8_t)(1 << 7))
 };
 //Extern global variables
 //Static global variables
